// frontend/src/backend.d.ts
export type Time = bigint;
// ...
export interface EmployeeRecord {
    internalId: bigint;
    status: string;
    leaveBalance: LeaveBalance;
    personalEmail?: string;
    designation: string;
    role: EmployeeRole;
    joiningDate: Time;
    businessEmail: string;
    employmentType: string;
    employeeId: string;
    department: string;
    phoneNumber: string;
}
// ...
export interface LeaveBalance {
    sick: bigint;
    annual: bigint;
    unpaid: bigint;
    casual: bigint;
}
export interface UserProfile {
    name: string;
    role: EmployeeRole;
    employeeId?: string;
}
export enum EmployeeRole {
    hrManager = "hrManager",
    manager = "manager",
    superAdmin = "superAdmin",
    employee = "employee",
    hrStaff = "hrStaff"
}